server/requirements.txt
fastapi
uvicorn[standard]
pyserial-asyncio
